Refactor caloriemanager, move timeframe picker, use decimal (#12)
* Delete unused code

* Create CalorieManagerUnitTests file

* Add tests for sumValueForDay

* Begin integrating environment into CalorieManager

* Update predicate tests

* Add subset with days

* Add healthSample function

* add date sorting test

* Rename HealthKitType

* WIP testing expected weights. Seems like it's not quite right

* Convert to decimal and comment out old code

* Fix tests

* Add copy and dropping

* Refactor getDays

* Add environment switcher for save

* Move TimePicker to bottom, might be laggy

* Small refactor

* Fix lag for picker view

* Add weightWasEstimated

---------

+++ Shared/Data/Weight.swift
@@ -8,16 +8,16 @@
 import Foundation
 
 struct Weight: Codable, HasDate {
-    var weight: Double
+    var weight: Decimal
     var date: Date
     
-    static func weightBetweenTwoWeights(date: Date, weight1: Weight?, weight2: Weight?) -> Double {
+    static func weightBetweenTwoWeights(date: Date, weight1: Weight?, weight2: Weight?) -> Decimal {
         guard let weight1 = weight1, let weight2 = weight2 else { return 0 }
         let days = Date.daysBetween(date1: weight1.date, date2: weight2.date)
-        guard days != 0 else { return Double(weight1.weight) }
-        let progress = Double(Date.daysBetween(date1: weight1.date, date2: date) ?? 1) / Double(days ?? 1)
-        let weightProgress = Double(weight2.weight - weight1.weight) * progress
-        let weight = Double(weight1.weight) + weightProgress
+        guard days != 0 else { return weight1.weight }
+        let progress = Decimal(Date.daysBetween(date1: weight1.date, date2: date) ?? 1) / Decimal(days ?? 1)
+        let weightProgress = (weight2.weight - weight1.weight) * progress
+        let weight = weight1.weight + weightProgress
         return weight
     }
